Refactor the button code
The code to make a choice through buttons is rewritten and the
previous code is deleted.

In the previous code we had a querySelector() and an eventListener
code for each button, where we manually added a value for the player
choice depending on each button.
Now we use a querySelectorAll() to select all the buttons and
get a value depending on their class.

On that way we use less code and is more readable.